refactor(node): split Node into package with base implementation and re-exports
Create `src/meridian/core/node/` with `__init__.py` and `base.py`, remove monolithic `node.py`, and preserve imports through package. Tests pass.
