+++ src/meridian/core/node/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .base import Node
+
+__all__ = ["Node"]
